Add user authentication with OAuth
Add an authentication system allowing users to sign in using either
Gitlab or Github. The about page is used to show a list of all
registered users.

Users has many links associated with them, each account link
holds the OAuth data for a provider, right now Gitlab and Github.
It is easy to extend this with more provider in the future, such
as Bitbucket for example.

Users are identified by their email which has a unique constraint
on it.

Database is changed to postgres for development and tests as well
to better reflect the production environment.

Code linting is added to CI tests and a rubocop configuration file
is created for improvements to do later. Code coverage is reported
to CodeClimate.