Import projects from Gitlab and Github
Allow users to list all their projects from Gitlab and Github and
select the projects to import into Appatite.

If the user has no projects active, they are redirected to the
setup page automatically.

When projects are deactivated they are not removed, just marked as
inactive, this will keep any future settings for the project, in
case the user reactivates them.

Fixes #7