Push out webhook changes to clients in real time
Whenever a webhook payload is recieved the changes are pushed out
to clients in real time, updating the web page in the browsers
without the need for a page refresh.

Also, remove the server command in Vagrantfile, leaving that up
to the user. This means that provisioning will succeeed as it will
actually reach a proper end.

Fixes #37