Disable CSRF protection for webhooks
Since the webhooks do not carry a CSRF token we have to disable
CSRF protection for those requests.

If this lower security too much there is always the ability to
register a secret, static token when creating the hook and using
it to verify the sender of the request. But for now I believe it
is not necessary as the webhooks only update non-sensitive information
such as build state on projects.

Fixes #40