Create a pricing page
Rename the `about` page to `pricing` and add description of the three
pricing tiers with their features and the price.

Style the navbar with a custom blue color and white text, and
replace the logged in user's name with their avatar instead. Similarly,
replace the Admin link with an icon (only non-mobile) and move it to
the right.

Fixes #26