Make users follow projects
Projects are now followable and users are able to be
followers. When a user activates a project, they follow
it.

When the owner deactivates a project, the webhook is
removed.

Editing of projects is restricted to owners and admins and
only admins are allowed to completely remove projects
from the database.

This fixes #35